refactor: align item generator class names with parent class (#2760)
