+++ streampipes-data-explorer/src/main/java/org/apache/streampipes/dataexplorer/query/writer/ConfiguredJsonOutputWriter.java
@@ -20,7 +20,7 @@ package org.apache.streampipes.dataexplorer.query.writer;
 
 import org.apache.streampipes.dataexplorer.param.ProvidedRestQueryParams;
 import org.apache.streampipes.dataexplorer.query.writer.item.ItemGenerator;
-import org.apache.streampipes.dataexplorer.query.writer.item.JsonItemWriter;
+import org.apache.streampipes.dataexplorer.query.writer.item.JsonItemGenerator;
 
 import com.google.gson.Gson;
 
@@ -33,11 +33,11 @@ public class ConfiguredJsonOutputWriter extends ConfiguredOutputWriter {
   private static final String BEGIN_ARRAY = "[";
   private static final String END_ARRAY = "]";
 
-  private final ItemGenerator jsonObjectWriter;
+  private final ItemGenerator itemGenerator;
 
   public ConfiguredJsonOutputWriter() {
     Gson gson = new Gson();
-    this.jsonObjectWriter = new JsonItemWriter(gson);
+    this.itemGenerator = new JsonItemGenerator(gson);
   }
 
   @Override
@@ -65,7 +65,7 @@ public class ConfiguredJsonOutputWriter extends ConfiguredOutputWriter {
       outputStream.write(toBytes(","));
     }
 
-    var item = jsonObjectWriter.createItem(row, columnNames);
+    var item = itemGenerator.createItem(row, columnNames);
     outputStream.write(toBytes(item));
   }
 }
