+++ streampipes-data-explorer/src/main/java/org/apache/streampipes/dataexplorer/query/writer/item/CsvItemGenerator.java
@@ -19,32 +19,19 @@
 
 package org.apache.streampipes.dataexplorer.query.writer.item;
 
-import com.google.gson.Gson;
+public class CsvItemGenerator extends ItemGenerator {
 
-public class JsonItemWriter extends ItemGenerator {
-
-  private static final String BEGIN_OBJECT = "{";
-  private static final String END_OBJECT = "}";
-
-  private final Gson gson;
-
-  public JsonItemWriter(Gson gson) {
-    super(COMMA_SEPARATOR);
-    this.gson = gson;
+  public CsvItemGenerator(String delimiter) {
+    super(delimiter);
   }
 
   @Override
-  protected String makeItemString(String key,
-                                  Object value) {
-    var stringValue = value != null ? gson.toJson(value) : null;
-    return "\""
-        + key
-        + "\": "
-        + stringValue;
+  protected String makeItemString(String key, Object value) {
+    return value != null ? value.toString() : "";
   }
 
   @Override
   protected String finalizeItem(String item) {
-    return BEGIN_OBJECT + item + END_OBJECT;
+    return item;
   }
 }
