+++ streampipes-data-explorer/src/test/java/org/apache/streampipes/dataexplorer/query/writer/item/TestItemGenerator.java
@@ -18,21 +18,19 @@
 
 package org.apache.streampipes.dataexplorer.query.writer.item;
 
-import com.google.gson.Gson;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.BeforeEach;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import java.util.Arrays;
+import java.util.List;
 
-public class TestJsonItemWriter extends TestItemWriter {
+public abstract class TestItemGenerator {
 
-  private static final String Expected = "{\"time\": 1668578077051,\"string\": \"test\",\"number\": 1}";
+  protected List<Object> row;
+  protected List<String> columns;
 
-  @Test
-  public void testJsonWriter() {
-    var writer = new JsonItemWriter(new Gson());
-
-    String result = writer.createItem(row, columns);
-
-    assertEquals(Expected, result);
+  @BeforeEach
+  public void before() {
+    this.row = Arrays.asList(1668578077051.0, "test", 1);
+    this.columns = Arrays.asList("time", "string", "number");
   }
 }
